requirements.txt
asgiref==3.3.2
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
